move reader_go18* to reader*
